+++ WGUCourseTracker/WGUCourseTracker/TermViewPage.xaml.cs
@@ -18,13 +18,14 @@ namespace WGUCourseTracker
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class TermViewPage : ContentPage
     {
-        Term term;
+        Term selectedTerm;
         Course course;
-
-        public TermViewPage()
+        ObservableCollection<Course> termCourses;
+        public TermViewPage(Term term)
         {
             InitializeComponent();
             term = (Term)mainStackLayout.BindingContext;
+            selectedTerm = term;          
             var selectedCourse = (Course)courseListView.SelectedItem;
 
         }
@@ -33,38 +34,34 @@ namespace WGUCourseTracker
             base.OnAppearing();
             courseListView.SelectedItem = null;
 
-
             using (SQLiteConnection con = new SQLiteConnection(App.DbLocation))
             {
-                //course = (Course)courseStackLayout.BindingContext;
-                term = (Term)mainStackLayout.BindingContext;
-                var termID = term.TermID;
+                selectedTerm = (Term)mainStackLayout.BindingContext;
+                var termID = selectedTerm.TermID;
 
                 con.CreateTable<Course>();
                 var courseTable = con.Table<Course>().ToList();
 
                 var courseList = (from course in courseTable
-                                  where course.TermID == term.TermID
-                                  select course);        
-               
+                                  where course.TermID == selectedTerm.TermID
+                                  select course);
+
+                termCourses = new ObservableCollection<Course>(courseList);
                 courseListView.ItemsSource = courseList;
             }
-
-
 
         }
         async void SaveTerm_Clicked(object sender, EventArgs e)
         {
-
             using (SQLiteConnection con = new SQLiteConnection(App.DbLocation))
             {
                 con.CreateTable<Term>();
 
                 if (DateCheck() && NullCheck())
                 {
-                    con.Update(term);
-                    await DisplayAlert("Success!", $"{term.TermName} Saved", "Ok");
-                    Page new1 = new TermListPage();
+                    con.Update(selectedTerm);
+                    await DisplayAlert("Success!", $"{selectedTerm.TermName} Saved", "Ok");
+                    Page new1 = new TermViewPage(selectedTerm) { BindingContext = selectedTerm as Term };
                     Navigation.InsertPageBefore(new1, this);
                     await Navigation.PopAsync();
                 }
@@ -77,9 +74,17 @@ namespace WGUCourseTracker
 
         async void NewCourse_Clicked(object sender, EventArgs e)
         {
-                var selectedTerm = (Term)mainStackLayout.BindingContext;
-
+            var selectedTerm = (Term)mainStackLayout.BindingContext;
+            if (termCourses.Count < 6)
+            {
                 await Navigation.PushAsync(new CourseAddPage { BindingContext = selectedTerm as Term });
+            }
+            else
+            {
+                await DisplayAlert("Error!", $"{selectedTerm.TermName} Has Maximum number of courses!", "Ok");
+            }
+            
+            
         }
 
         async void EditCourse_Clicked(object sender, EventArgs e)
@@ -89,15 +94,13 @@ namespace WGUCourseTracker
             if (selectedCourse != null)
             {
                 await Navigation.PushAsync(new CourseViewPage(course) { BindingContext = selectedCourse as Course });
-                //await Navigation.PushAsync(new CourseViewPage { BindingContext = selectedCourse as Course });
             }
             else
             {
                 await DisplayAlert("Error!", "Please Select a Course.", "Ok");
             }
-            
-
         }
+
         async void DeleteButton_Clicked(object sender, EventArgs e)
         {
             var selectedCourse = (Course)courseListView.SelectedItem;
@@ -106,14 +109,23 @@ namespace WGUCourseTracker
             {
                 using (SQLiteConnection con = new SQLiteConnection(App.DbLocation))
                 {
-                    bool deleteCourse = await DisplayAlert("Confirm!", $"Do you wish to delete the '{selectedCourse.CourseName}' course?", "Yes", "No");
-                    if (deleteCourse)
+                    var courseAssessments = con.Query<Course>($"Select * from Assessment where CourseID = {selectedCourse.CourseID}");
+                    
+                    if(courseAssessments.Count == 0)
                     {
-                        con.Execute($"Delete from Course Where {selectedCourse.CourseID} = CourseID ");
-                        Page new1 = new TermViewPage();
-                        new1.BindingContext = term as Term;
-                        Navigation.InsertPageBefore(new1, this);
-                        await Navigation.PopAsync();
+                        bool deleteCourse = await DisplayAlert("Confirm!", $"Do you wish to delete the '{selectedCourse.CourseName}' course?", "Yes", "No");
+                        if (deleteCourse)
+                        {
+                            con.Execute($"Delete from Course Where {selectedCourse.CourseID} = CourseID ");
+                            Page new1 = new TermViewPage(selectedTerm) { BindingContext = selectedTerm as Term };
+                            //new1.BindingContext = term as Term;
+                            Navigation.InsertPageBefore(new1, this);
+                            await Navigation.PopAsync();
+                        }
+                    }
+                    else
+                    {
+                        await DisplayAlert("Unable to Delete!", $"Course has {courseAssessments.Count} assessment(s) associated with it!", "Ok");
                     }
 
                 }
@@ -121,9 +133,9 @@ namespace WGUCourseTracker
             else
             {
                 await DisplayAlert("Error!", "Please Select a Course to Remove!", "Ok");
-            }
-            
+            }           
         }
+
         private bool NullCheck()
         {
             if (!String.IsNullOrEmpty(termEntry.Text))
@@ -133,6 +145,7 @@ namespace WGUCourseTracker
             DisplayAlert("Error!", "Please fill in all information!", "Ok");
             return false;
         }
+
         private bool DateCheck()
         {
             if (startDatePicker.Date < endDatePicker.Date)
@@ -142,8 +155,6 @@ namespace WGUCourseTracker
             DisplayAlert("Error!", "Start Date must be greater than End Date!", "OK");
             return false;
         }
-
-
 
     }
 }
